remove aggregates dir and rename domain models to readModels
