+++ src/db/models/DBBadge.ts
@@ -4,9 +4,7 @@ export type BadgeSlug = string;
 export interface BadgeRelation {
   badge_id: DBBadge["id"];
 }
-// table name: badges
-export interface DBBadge extends DBDatedData {
-  id: number;
+export interface DBInsertBadge {
   slug: BadgeSlug;
   name: string;
 }
