+++ src/db/models/app/DBCategory.ts
@@ -1,5 +1,5 @@
 import { DBDatedData } from "@db/models/app/DBDatedData";
-import { AppCategoryName } from "@domain/models/app/Category";
+import { AppCategoryName } from "@domain/readModels/app/Category";
 
 export interface CategoryNameRelation {
   category: DBCategory["name"];
