+++ src/db/models/app/DBProject.ts
@@ -1,8 +1,8 @@
 import { DBUser, UserRelation } from "./DBUser";
 import { DBDatedData } from "./DBDatedData";
 import { VersionRelation } from "./DBVersion";
-import { AppCategoryName } from "@domain/models/app/Category";
-import { ProjectSlug } from "@domain/models/app/Project";
+import { AppCategoryName } from "@domain/readModels/app/Category";
+import { ProjectSlug } from "@domain/readModels/app/Project";
 
 export interface DBInsertProject
   extends Partial<VersionRelation>, // The Latest DBVersion
