+++ src/domain/readModels/ProjectStatusOnBadge.ts
@@ -1,6 +1,6 @@
-import { DatedData } from "@domain/models/app/DatedData";
+import { DatedData } from "@domain/readModels/app/DatedData";
 import { BadgeRelation } from "./Badge";
-import { ProjectStatusName } from "@domain/models/app/Project";
+import { ProjectStatusName } from "@domain/readModels/app/Project";
 
 export interface ProjectStatusOnBadge extends BadgeRelation, DatedData {
   status: ProjectStatusName; // Status for this project for this particular badge
