+++ src/domain/readModels/app/Category.ts
