+++ src/domain/readModels/app/DatedData.ts
