+++ src/domain/readModels/app/FileMetadata.ts
