+++ src/domain/readModels/app/VoteFromUser.ts
